Move toCdaJson logic to utils
